+++ lib/v8.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const v8binding = process.binding('v8');
-const smalloc = require('smalloc');
+const smalloc = require('internal/smalloc');
 
 const heapStatisticsBuffer =
     smalloc.alloc(v8binding.kHeapStatisticsBufferLength,
